Refactor exception handling and code cleanup.
Simplified exception handling with multi-catch for `ExecutionException` and `InterruptedException` in `SimpleReelGenerator`. Removed redundant type casting in `VirtualReel`, replaced explicit generic type with diamond operator in `ArrayList`, and added an unused warning suppression in `VirtualReelBuilder`. These changes enhance code readability and maintainability.

+++ src/main/java/wtd/slotsengine/slots/machines/reels/VirtualReel.java
@@ -44,7 +44,7 @@ public class VirtualReel implements IReel {
         int len = data.size();
         byte[] result = new byte[len];
         for (int i = 0; i < len; i++) {
-            result[i] = data.get(i).byteValue();
+            result[i] = data.get(i);
         }
         return result;
     }
